requirements.txt
# Core dependencies
pytesseract
Pillow
pytest

# AI Analysis (optional)
llama-cpp-python

# Video Analysis (optional)
opencv-python

# Translation (optional)
langdetect
googletrans==4.0.0-rc1
